+++ R/import_list.R
@@ -1,34 +1,42 @@
 #' @title Import list of data frames
 #' @description Use \code{\link{import}} to import a list of data frames from a vector of file names or from a multi-object file (Excel workbook, .Rdata file, zip directory, or HTML file)
 #' @param file A character string containing a single file name for a multi-object file (e.g., Excel workbook, zip directory, or HTML file), or a vector of file paths for multiple files to be imported.
+#' @template setclass
 #' @param which If \code{file} is a single file path, this specifies which objects should be extracted (passed to \code{\link{import}}'s \code{which} argument). Ignored otherwise.
+#' @param rbind A logical indicating whether to pass
 #' @param \dots Additional arguments passed to \code{\link{import}}. Behavior may be unexpected if files are of different formats.
-#' @return A list of a data frames.
+#' @return If \code{rbind=FALSE} (the default), a list of a data frames. Otherwise, that list is passed to \code{\link[data.table]{rbindlist}} with \code{fill = TRUE} and returns a data frame object of class set by the \code{setclass} argument; if this operation fails, the list is returned.
 #' @examples
 #' library('datasets')
-#' export(list(iris = iris, mtcars = mtcars), "data.xlsx")
+#' export(list(mtcars1 = mtcars[1:10,], 
+#'             mtcars2 = mtcars[11:20,],
+#'             mtcars2 = mtcars[21:32,]), "mtcars.xlsx")
 #' 
 #' # import a single file from multi-object workbook
-#' str(import("data.xlsx", which = "mtcars"))
+#' str(import("mtcars.xlsx", which = "mtcars1"))
 #' 
 #' # import all worksheets
-#' str(import_list("data.xlsx"))
+#' str(import_list("mtcars.xlsx"), 1)
+#' 
+#' # import and rbind all worksheets
+#' mtcars2 <- import_list("mtcars.xlsx", rbind = TRUE)
+#' all.equal(mtcars2, mtcars, check.attributes = FALSE)
 #' 
 #' # import multiple files
 #' export(mtcars, "mtcars.csv")
 #' export(mtcars, "iris.csv")
-#' str(import_list(c("mtcars.csv", "iris.csv")))
+#' str(import_list(dir(pattern = "csv$")), 1)
 #' 
 #' # cleanup
-#' unlink("data.xlsx")
+#' unlink("mtcars.xlsx")
 #' unlink("mtcars.csv")
 #' unlink("iris.csv")
 #' 
 #' @seealso \code{\link{import}}
 #' @export
-import_list <- function(file, which, ...) {
+import_list <- function(file, setclass, which, rbind = FALSE, ...) {
     if (length(file) > 1) {
-        lapply(file, import, ...)
+        x <- lapply(file, import, ...)
     } else {
         if (get_ext(file) == "rdata") {
             e <- new.env()
@@ -37,10 +45,10 @@ import_list <- function(file, which, ...) {
         }
         if (missing(which)) {
             if (get_ext(file) == "html") {
-                requireNamespace("xml2")
+                requireNamespace("xml2", quietly = TRUE)
                 which <- seq_along(xml2::xml_find_all(xml2::read_html(unclass(file)), ".//table"))
             } else if (get_ext(file) %in% c("xls","xlsx")) {
-                requireNamespace("readxl")
+                requireNamespace("readxl", quietly = TRUE)
                 which <- seq_along(readxl::excel_sheets(path = file))
             } else if (get_ext(file) %in% c("zip")) {
                 which <- seq_len(nrow(utils::unzip(file, list = TRUE)))
@@ -48,6 +56,32 @@ import_list <- function(file, which, ...) {
                 which <- 1
             }
         }
-        lapply(which, function(w) import(file, which = w, ...))
+        x <- lapply(which, function(w) import(file, which = w, ...))
     }
+    
+    # optionally rbind
+    if (isTRUE(rbind)) {
+        if (length(x) == 1) {
+            x <- x[[1L]]
+        } else {
+            x2 <- try(data.table::rbindlist(x, fill = TRUE), silent = TRUE)
+            if (inherits(x2, "try-error")) {
+                warnings("Attempt to rbindlist() the data did not succeed. List returned instead.")
+                return(x)
+            } else {
+                x <- x2
+            }
+        }
+        # set class
+        if (missing(setclass)) {
+            return(set_class(x))
+        }
+        a <- list(...)
+        if ("data.table" %in% names(a) && isTRUE(a[["data.table"]])) {
+            setclass <- "data.table"
+        }
+        x <- set_class(x, class = setclass)
+    }
+    
+    return(x)
 }