+++ man-roxygen/setclass.R
@@ -1,0 +1,1 @@
+#' @param setclass An optional character vector specifying one or more classes to set on the import. By default, all the return object is always a \dQuote{data.frame}. Allowed values for this might be \dQuote{tbl_df}, \dQuote{tbl}, or \dQuote{tibble} (if using dplyr) or \dQuote{data.table} (if using data.table). Other values are ignored such that a data.frame is returned.